Task 4/Automated Resume Screener/requirements.txt
Flask>=2.0
Flask-Mail>=0.9.1
pdfplumber>=0.7.6
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.0